requirements.txt
certifi==2020.6.20
Pillow==8.2.0
pyinstaller==4.3
PyPDF2==1.26.0
